requirements.txt
streamlit==1.30.0
streamlit-webrtc==0.57.0
mediapipe==1.15.0
opencv-python==4.8.0
numpy==1.26.0
av==10.0.0
Pillow==10.0.0